refactor: centralize OGP and meta tag configuration into a `useOgpHead` composable for page components.
